requirements.txt
streamlit>=1.22.0
pandas>=1.5.0
numpy>=1.23.0
yfinance>=0.2.18
matplotlib>=3.7.0
plotly>=5.14.0
scikit-learn>=1.2.0
torch>=2.0.0
transformers>=4.35.0
requests>=2.28.0
python-dotenv>=1.0.0
accelerate>=0.22.0
bitsandbytes>=0.41.0
scipy>=1.10.0